Changed type of device ID to long

+++ src/io/github/vladast/avrcommunicator/db/dao/SessionDAO.java
@@ -33,13 +33,13 @@ public class SessionDAO extends EventRecorderDAO {
 		);
 	*/
 	
-	private int mIdDevice;
+	private long mIdDevice;
 	
-	public int getIdDevice() {
+	public long getIdDevice() {
 		return mIdDevice;
 	}
 	
-	public void setIdDevice(int idDevice) {
+	public void setIdDevice(long idDevice) {
 		mIdDevice = idDevice;
 	}
